TemplateApp is main base for managing dependencies as per yarn monorepo principle, Framework to be moved out of projects folder
