+++ .gitignore
@@ -155,7 +155,6 @@ lerna-debug.log*
 .DS_Store
 Thumbs.db
 
-# Custom rules for projects directory
+# ignore all projects except TemplateApp
 /projects/*
-!/projects/Framework/
 !/projects/TemplateApp/
