+++ Framework/Common/Values.js
