+++ Framework/Managers/LocalDataManager.js
