+++ Framework/UI/SearchBar.js
