+++ package.json
@@ -1,39 +1,11 @@
 {
-  "name": "expo-app-framework",
   "private": true,
-  "dependencies": {
-    "@expo/webpack-config": "^19.0.0",
-    "@react-native-async-storage/async-storage": "1.18.2",
-    "@react-native-picker/picker": "2.4.10",
-    "@react-navigation/native": "^6.1.9",
-    "@react-navigation/native-stack": "^6.9.17",
-    "axios": "^1.6.1",
-    "expo": "^49.0.17",
-    "expo-build-properties": "~0.8.3",
-    "expo-status-bar": "~1.6.0",
-    "form-data": "^4.0.0",
-    "lodash": "^4.17.21",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "0.72.6",
-    "react-native-big-list": "^1.6.1",
-    "react-native-collapsible": "^1.6.1",
-    "react-native-gesture-handler": "~2.12.0",
-    "react-native-get-random-values": "~1.9.0",
-    "react-native-pager-view": "6.2.0",
-    "react-native-paper": "^5.11.1",
-    "react-native-popup-menu": "^0.16.1",
-    "react-native-safe-area-context": "4.6.3",
-    "react-native-screens": "~3.22.0",
-    "react-native-tab-view": "^3.5.2",
-    "react-native-web": "~0.19.6",
-    "uuid": "^9.0.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "@faker-js/faker": "^8.2.0"
-  },
+  "name": "expo-app-framework",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
   "workspaces": [
-    "projects/*"
+    "projects/*",
+    "Framework"
   ]
 }
