+++ projects/TemplateApp/Pages/SampleMenusPage.js
@@ -1,17 +1,17 @@
 import React, { useContext, useState, useEffect, useCallback, useRef } from 'react';
 import { View, Keyboard } from 'react-native';
-import Styles from '@expo-app-framework/framework/Common/Styles';
+import Styles from '../../../Framework/Common/Styles';
 // UI
 import {
   useTheme, Text, Card, Button, Appbar,
   TouchableRipple, Searchbar, IconButton, FAB, Portal, Divider, Snackbar
 } from 'react-native-paper';
-import Dialog from '@expo-app-framework/framework/UI/Dialog';
-import Picker from '@expo-app-framework/framework/UI/Picker';
-import DropdownMenu from '@expo-app-framework/framework/UI/DropdownMenu';
-import DropdownCheckMenu from '@expo-app-framework/framework/UI/DropdownCheckMenu';
+import Dialog from '../../../Framework/UI/Dialog';
+import Picker from '../../../Framework/UI/Picker';
+import DropdownMenu from '../../../Framework/UI/DropdownMenu';
+import DropdownCheckMenu from '../../../Framework/UI/DropdownCheckMenu';
 // data
-import { LocalDataContext } from '@expo-app-framework/framework/Contexts/LocalDataContext';
+import { LocalDataContext } from '../../../Framework/Contexts/LocalDataContext';
 // const
 const PICKER_ITEM_LIST = [
   { label: 'Red', value: 'red' },
