+++ projects/TemplateApp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@expo-app-framework/template-app",
+  "name": "template-app",
   "version": "1.0.0",
   "main": "./AppEntry.js",
   "scripts": {
@@ -8,8 +8,36 @@
     "ios": "expo start --ios",
     "web": "expo start --web"
   },
-  "private": true,
   "dependencies": {
-    "@expo-app-framework/framework": "1.0.0"
+    "@expo/metro-runtime": "~3.2.1",
+    "@react-native-async-storage/async-storage": "1.23.1",
+    "@react-native-picker/picker": "2.7.5",
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/native-stack": "^6.9.26",
+    "axios": "^1.7.2",
+    "expo": "~51.0.8",
+    "expo-build-properties": "~0.12.1",
+    "expo-status-bar": "~1.12.1",
+    "form-data": "^4.0.0",
+    "lodash": "^4.17.21",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-native": "0.74.1",
+    "react-native-big-list": "^1.6.1",
+    "react-native-collapsible": "^1.6.1",
+    "react-native-gesture-handler": "~2.16.1",
+    "react-native-get-random-values": "~1.11.0",
+    "react-native-pager-view": "6.3.0",
+    "react-native-paper": "^5.12.3",
+    "react-native-popup-menu": "^0.16.1",
+    "react-native-safe-area-context": "4.10.1",
+    "react-native-screens": "3.31.1",
+    "react-native-tab-view": "^3.5.2",
+    "react-native-web": "~0.19.10",
+    "uuid": "^9.0.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.20.0",
+    "@faker-js/faker": "^8.4.1"
   }
 }
